@@ -5,14 +5,12 @@ description: >-
 author: Daco Harkes <dacoharkes@google.com>, Samir Jindel <sjindel@google.com>
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
-  flutter: '>=1.5.9-pre.270'
+  sdk: ^2.12.0-0
 
 dependencies:
   flutter:
     sdk: flutter
-  ffi: ^0.1.3
-
+  ffi: ^0.2.0-nullsafety.1
 dev_dependencies:
   test: ^1.5.3
   flutter_test:
