@@ -1,9 +1,7 @@
 class SqliteConfig {
   const SqliteConfig({this.documentPath});
 
-  final String documentPath;
+  final String? documentPath;
 
-  static const SqliteConfig builtin = SqliteConfig(
-    documentPath: null,
-  );
+  static const builtin = SqliteConfig();
 }
